Moving Fetch,Remove Books to Hook File
