+++ src/hooks/useBook.ts
@@ -1,6 +1,9 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
+// Types
+import { Book } from '../types/'
+
 type ValueDate = Date;
 type Value = ValueDate[];
 
@@ -8,6 +11,9 @@ export const useBook = () => {
   const [isBooking, setBooking] = useState(false);
   const [value, onChange] = useState<Value>([new Date()]);
   const [error, setError] = useState(false);
+
+  const [books, setBooks] = useState<Book[]>([]);
+  const [removing, setRemoving] = useState({ id: '', active: false });
 
   const navigate = useNavigate();
 
@@ -23,6 +29,33 @@ export const useBook = () => {
     const days = getDays(value[0], value[1]);
 
     return price * days;
+  }
+
+  const fetchBooks = async () => {
+    try {
+      const data = await (await fetch(`${import.meta.env.VITE_API_URL}/books`)).json();
+      setBooks(data.books);
+    } catch (err) {
+      console.log(err);
+    }
+  }
+
+  const removeBook = async (id: string) => {
+    try {
+      setRemoving({ id, active: true });
+  
+      await fetch(`${import.meta.env.VITE_API_URL}/book`, {
+        method: 'DELETE',
+        body: JSON.stringify({
+          id,
+        })
+      });
+      
+      setRemoving({ id: '', active: false });
+      setBooks(books.filter(book => book.id !== id))
+    } catch (err) {
+      setRemoving({ id: '', active: false });
+    }
   }
 
   const createBook = async (id: string, price: number) => {
@@ -52,11 +85,15 @@ export const useBook = () => {
     isBooking,
     value,
     error,
+    removing,
+    books,
     setBooking,
     onChange,
     setError,
     getCurrentPrice,
     getDays,
-    createBook
+    createBook,
+    fetchBooks,
+    removeBook,
   }
 }
